hardware/printer/requirements.txt
python-escpos>=3.1
Pillow>=10.0
